# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hatim API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v2/groups": {
      "get": {
        "tags": [
          "Circles"
        ],
        "summary": "Groups",
        "operationId": "groups_api_v2_groups_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CircleSummary"
                  },
                  "type": "array",
                  "title": "Response Groups Api V2 Groups Get"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Circles"
        ],
        "summary": "Create Circle",
        "operationId": "create_circle_api_v2_groups_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CircleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CircleView"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v2/groups/{circle_id}": {
      "get": {
        "tags": [
          "Circles"
        ],
        "summary": "Group",
        "operationId": "group_api_v2_groups__circle_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "circle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Circle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CircleView"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/groups/{circle_id}/title": {
      "put": {
        "tags": [
          "Circles"
        ],
        "summary": "Rename",
        "operationId": "rename_api_v2_groups__circle_id__title_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "circle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Circle Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TitleChange"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CircleView"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/invites/{code}": {
      "get": {
        "tags": [
          "Circles"
        ],
        "summary": "Invite",
        "operationId": "invite_api_v2_invites__code__get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicCircle"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/groups": {
      "post": {
        "summary": "Create Group",
        "operationId": "create_group_api_groups_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroup"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupCreated"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Plans"
        ],
        "summary": "Account Plans",
        "operationId": "account_plans_api_groups_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanSummary"
                  },
                  "title": "Response Account Plans Api Groups Get"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{group_id}/account": {
      "put": {
        "tags": [
          "Plans"
        ],
        "summary": "Save To Account",
        "operationId": "save_to_account_api_groups__group_id__account_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveToAccount"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupView"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{group_id}/title": {
      "put": {
        "tags": [
          "Plans"
        ],
        "summary": "Rename Plan",
        "operationId": "rename_plan_api_groups__group_id__title_put",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TitleChange"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupView"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{group_id}": {
      "delete": {
        "tags": [
          "Plans"
        ],
        "summary": "Delete Plan",
        "operationId": "delete_plan_api_groups__group_id__delete",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Acknowledged"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get Group",
        "operationId": "get_group_api_groups__group_id__get",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupView"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{group_id}/settings": {
      "put": {
        "summary": "Update Settings",
        "operationId": "update_settings_api_groups__group_id__settings_put",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanSettingsChange"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupView"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{group_id}/profile": {
      "put": {
        "summary": "Update Organizer",
        "operationId": "update_organizer_api_groups__group_id__profile_put",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Preferences"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupView"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{group_id}/members/{member_id}": {
      "delete": {
        "summary": "Remove Member",
        "operationId": "remove_member_api_groups__group_id__members__member_id__delete",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Id"
            }
          },
          {
            "name": "member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Member Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupView"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/invites/{code}": {
      "get": {
        "summary": "Invite",
        "operationId": "invite_api_invites__code__get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteView"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/invites/{code}/members": {
      "post": {
        "summary": "Join Group",
        "operationId": "join_group_api_invites__code__members_post",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Preferences"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberCreated"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/invites/{code}/me": {
      "get": {
        "summary": "Get My Preferences",
        "operationId": "get_my_preferences_api_invites__code__me_get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Member"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update My Preferences",
        "operationId": "update_my_preferences_api_invites__code__me_put",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Preferences"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Member"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Acknowledged": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Acknowledged"
      },
...
      "CircleCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "Title"
          },
          "preferences": {
            "$ref": "#/components/schemas/Preferences"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "preferences"
        ],
        "title": "CircleCreate"
      },
...
      "CircleSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "member_count": {
            "type": "integer",
            "title": "Member Count"
          },
          "pinned": {
            "type": "boolean",
            "title": "Pinned"
          },
          "is_owner": {
            "type": "boolean",
            "title": "Is Owner"
          },
          "archived": {
            "type": "boolean",
            "title": "Archived"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "member_count",
          "pinned",
          "is_owner",
          "archived"
        ],
        "title": "CircleSummary"
      },
      "CircleView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "invite_code": {
            "type": "string",
            "title": "Invite Code"
          },
          "is_owner": {
            "type": "boolean",
            "title": "Is Owner"
          },
          "archived": {
            "type": "boolean",
            "title": "Archived"
          },
          "me": {
            "$ref": "#/components/schemas/CircleMember"
          },
          "pinned": {
            "type": "boolean",
            "title": "Pinned"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/CircleMember"
            },
            "type": "array",
            "title": "Members"
          },
          "outings": {
            "items": {
              "$ref": "#/components/schemas/OutingSummary"
            },
            "type": "array",
            "title": "Outings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "invite_code",
          "is_owner",
          "archived",
          "me",
          "pinned",
          "members",
          "outings"
        ],
        "title": "CircleView"
      },
      "CreateGroup": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "Title",
            "default": "لَمّتنا في الرياض"
          },
          "preferences": {
            "$ref": "#/components/schemas/Preferences"
          },
          "settings": {
            "$ref": "#/components/schemas/Settings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "preferences"
        ],
        "title": "CreateGroup"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
...
      "GroupCreated": {
        "properties": {
          "organizer_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organizer Token"
          },
          "group": {
            "$ref": "#/components/schemas/GroupView"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "organizer_token",
          "group"
        ],
        "title": "GroupCreated"
      },
      "GroupView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "owner_account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Account Id"
          },
          "invite_code": {
            "type": "string",
            "title": "Invite Code"
          },
          "settings": {
            "$ref": "#/components/schemas/Settings"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/Member"
            },
            "type": "array",
            "title": "Members"
          },
          "plan": {
            "$ref": "#/components/schemas/Plan"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "invite_code",
          "settings",
          "members",
          "plan"
        ],
        "title": "GroupView"
      },
...
      "MemberCreated": {
        "properties": {
          "member_token": {
            "type": "string",
            "title": "Member Token"
          },
          "member": {
            "$ref": "#/components/schemas/Member"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "member_token",
          "member"
        ],
        "title": "MemberCreated"
      },
...
      "PlanSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "slots": {
            "type": "integer",
            "title": "Slots"
          },
          "consumed": {
            "type": "integer",
            "title": "Consumed"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "slots",
          "consumed"
        ],
        "title": "PlanSummary"
      },
...
      "PublicCircle": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "member_count": {
            "type": "integer",
            "title": "Member Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "member_count"
        ],
        "title": "PublicCircle"
      },
...
      "SaveToAccount": {
        "properties": {
          "owner_token": {
            "type": "string",
            "format": "password",
            "title": "Owner Token",
            "writeOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "owner_token"
        ],
        "title": "SaveToAccount"
      },
...
      "TitleChange": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "Title"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title"
        ],
        "title": "TitleChange"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}